% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCoveragePlot.R
\name{CompareCoveragePlot}
\alias{CompareCoveragePlot}
\title{Plot Comparing Coverage of ChIP-Seq Data}
\usage{
CompareCoveragePlot(chipset1, chipset2, chromosome)
}
\arguments{
\item{chipset1}{A GRanges object representing a set of ChIPseq data.}

\item{chipset2}{Another GRanges object representing a set of ChIPseq
data.}
}
\value{
Returns nothing, generating two coverage plots.
}
\description{
Plots the coverages of two sets of ChIP-Seq Data for the same chromosome.
This function requires some time to run.
}
\examples{
CompareCoveragePlot(cstest$gfp, cstest$ctcf)

}
\references{
Lawrence M, Huber W, Pagès H, Aboyoun P, Carlson M, Gentleman R, Morgan M, Carey V (2013). “Software for Computing and Annotating Genomic Ranges.” PLoS Computational Biology, 9. doi: 10.1371/journal.pcbi.1003118, http://www.ploscompbiol.org/article/info\%3Adoi\%2F10.1371\%2Fjournal.pcbi.1003118
}
