% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cstest}
\alias{cstest}
\title{A test ChIP-Seq dataset}
\format{
A GRanges object from 3 chromosomes, with 2 lanes:
\describe{
\item{cstest$gfp}{GFP pull-down in mouse.}
\item{cstest$ctcf}{CTCF pull-down in mouse.}
}
}
\source{
Short Read Archive, GEO accession number GSM288351 http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM288351
}
\usage{
cstest
}
\description{
A part of ChIP-Seq dataset obtained from the Short-Read Archive.
}
\examples{
\dontrun{
 cstest
}
}
\keyword{datasets}
