% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeOverlapPlot.R
\name{RangeOverlapPlot}
\alias{RangeOverlapPlot}
\title{Plot Overlap in Sequence Ranges}
\usage{
RangeOverlapPlot(chipset1, chipset2)
}
\arguments{
\item{chipset1}{A GRanges object representing a set of ChIPseq data.}

\item{chipset2}{Another GRanges object representing a set of ChIPseq
data.}
}
\value{
Returns a plot of ranges of sequence that have been identified
in both sets of data
}
\description{
Plots information criteria values given numeric information criteria
values.
}
\examples{
RangeOverlapPlot(cstest$gfp, cstest$ctcf)

}
