% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcommonTFfinder.R
\name{runcommonTFfinder}
\alias{runcommonTFfinder}
\title{Launch Shiny App for commonTFfinder}
\usage{
runcommonTFfinder()
}
\value{
No return value but open up a Shiny page.
}
\description{
This function launches the Shiny app for commonTFfinder.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
commonTFfinder::runcommonTFfinder()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
