% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PossibleBindingSites.R
\name{OverlappingSites}
\alias{OverlappingSites}
\title{Find Overlap in Sequence Ranges}
\usage{
OverlappingSites(chipset1, chipset2)
}
\arguments{
\item{chipset1}{A GRanges object representing a set of ChIPseq data.}

\item{chipset2}{Another GRanges object representing a set of ChIPseq
data.}
}
\value{
Returns a plot of ranges of sequence that have been identified
in both sets of data
}
\description{
Plots information criteria values given numeric information criteria
values.
Developmental Notes:
Use these packages once figure out problem
importFrom BSgenome.Mmusculus.UCSC.mm8 BSgenome.Mmusculus.UCSC.mm8
importFrom BSgenome.Hsapiens.UCSC.hg19 BSgenome.Hsapiens.UCSC.hg19
importFrom BSgenome.Hsapiens.UCSC.hg38 BSgenome.Hsapiens.UCSC.hg38
importFrom BSgenome.Mmusculus.UCSC.mm9 BSgenome.Mmusculus.UCSC.mm9
importFrom BSgenome.Mmusculus.UCSC.mm10 BSgenome.Mmusculus.UCSC.mm10
}
\examples{


}
\references{

}
